@@ -1,6 +1,6 @@
 package com.example.ajeshpai.androidanimatios.PhysicsBasedAnimations
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.example.ajeshpai.androidanimatios.R
 import kotlinx.android.synthetic.main.activity_main.*
