@@ -1,14 +1,14 @@
 package com.example.ajeshpai.androidanimatios.PhysicsBasedAnimations
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.support.animation.FloatPropertyCompat
-import android.support.animation.SpringAnimation
-import android.support.animation.SpringForce
 import android.util.Log
 import android.view.MotionEvent
 import android.view.View
 import android.widget.ImageView
+import androidx.dynamicanimation.animation.FloatPropertyCompat
+import androidx.dynamicanimation.animation.SpringAnimation
+import androidx.dynamicanimation.animation.SpringForce
 import com.example.ajeshpai.androidanimatios.R
 
 class SpringAnimationActivity : AppCompatActivity() {
