@@ -1,8 +1,8 @@
 package com.example.ajeshpai.androidanimatios.transitions
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.support.v7.widget.LinearLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
 import android.view.View
 import com.example.ajeshpai.androidanimatios.R
 import kotlinx.android.synthetic.main.search_view.*
